Pins pip, setuptools, pip-tools
The goal is to provide recent and predictable versions of pip and
setuptools. Also upgraded the version of pip-tools without it,

Adds --allow-unsafe in app-code deps

The pip-compile command *must* use --allow-unsafe in order to pin
seutptools and/or pip. We now do so. Adding this flag worked fine during
local invocation, i.e. running `pip-compile` on the host machine
(plus a few apt deps such as apache2-dev, enchant, etc),
but fails via the containerized make target with:

    Collecting pathlib2>=2.2.0; python_version < "3.6"
    ERROR: In --require-hashes mode, all requirements must have their versions pinned with ==. These do not:
        pathlib2>=2.2.0; python_version < "3.6" from https://files.pythonhosted.org/packages/e9/45/9c82d3666af4ef9f221cbb954e1d77ddbb513faf552aea6df5f37f1a4859/pathlib2-2.3.5-py2.py3-none-any.whl#sha256=0ec8205a157c80d7acc301c0b18fbd5d44fe655968f5d947b6ecef5290fc35db (from pytest==3.10->-r requirements/python3/test-requirements.txt (line 162))

It's possible a new version of pytest would resolve,
haven't tried that yet.